+++ webapp/src/App.css
@@ -14,18 +14,18 @@
 }
 
 .App-header {
-  background-color: #282c34;
+  background-color: #000; /* Changed to black */
   min-height: 100vh;
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
   font-size: calc(10px + 2vmin);
-  color: white;
+  color: #fff; /* Changed to white */
 }
 
 .App-link {
-  color: #61dafb;
+  color: #fff; /* Changed to white */
 }
 
 @keyframes App-logo-spin {
