+++ webapp/src/UsePages/Chef.css
@@ -1,0 +1,71 @@
+.Inc {
+    border-radius: 5px;
+    background: #000; /* Changed to black */
+    color: #fff; /* Changed to white */
+    transition: background-color 0.2s ease;
+  }
+  
+  .Inc:hover {
+    background-color: #555; /* Darker gray on hover */
+  }
+  
+  .Dec {
+    border-radius: 5px;
+    background: #000; /* Changed to black */
+    color: #fff; /* Changed to white */
+    transition: background-color 0.2s ease; 
+  }
+  
+  .Dec:hover {
+    background-color: #555; /* Darker gray on hover */
+  }
+  
+  /* Main container styling */
+  .Chef-container {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-between;
+    padding: 20px;
+  }
+  
+  /* Finalized order table styling */
+  .Finalized {
+    width: 28%;
+    margin-right: 20px;
+    padding: 20px;
+    border: 1px solid #000; /* Changed to black */
+    border-radius: 8px;
+    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
+    background-color: #f8f9fa; /* Light gray background */
+  }
+  
+  .Finalized h3 {
+    margin-bottom: 20px;
+    text-align: center;
+    font-family: 'Ubuntu', sans-serif; /* Use Ubuntu font */
+    font-weight: 600; /* Bold font for emphasis */
+    color: #000; /* Changed to black */
+  }
+  
+  .Finalized button {
+    width: 100%;
+    padding: 10px;
+    border: 1px solid #000; /* Changed to black */
+    background-color: #000; /* Changed to black */
+    color: #fff; /* Changed to white */
+    border-radius: 5px;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+  }
+  
+  .Finalized button:hover {
+    background-color: #555; /* Darker gray on hover */
+  }
+  
+  /* Button container styling */
+  .Button-container {
+    display: flex;
+    justify-content: space-around;
+    margin-top: 20px;
+  }
+
