+++ webapp/src/UsePages/Waiter.css
@@ -1,165 +1,162 @@
-
-/* Header styling */
-.Header h1 {
-  text-align: center; /* Center the text horizontally */
-  margin-bottom: 30px; /* More space below the header */
-  font-family: 'Ubuntu', sans-serif; /* Use Ubuntu font */
-  font-size: 32px; /* Increased font size for better visibility */
-  font-weight: 600; /* Slightly bolder font for emphasis */
-  color: #333; /* Dark color for text for better contrast */
-  border-bottom: 2px solid #28a745; /* Green bottom border for visual separation */
-  padding: 5px; /* Space between the text and the border */
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1); /* Subtle shadow for depth */
-  background-color: rgb(254, 255, 169); /* Light background color for a clean look */
-  border-radius: 5px; /* Slightly rounded corners */
-  max-width: 400px; /* Set a maximum width for the header */
-  margin: 0 auto; /* Center the header */
-}
-.icon {
-  width: 30px; 
-  height: 30px; 
-  margin-left: 10px; 
-  vertical-align: middle;
-  transition: transform 0.2s;
-}
-
-.Inc {
-  border-radius: 5px;
-  background: rgb(250, 151, 1);
-  color: #fff; 
-  transition: background-color 0.2s ease;
-}
-
-.Inc:hover {
-  background-color: #333; 
-}
-
-.Dec {
-  border-radius: 5px;
-  background: rgb(197, 0, 174);
-  color: #fff; 
-  transition: background-color 0.2s ease; 
-}
-
-.Dec:hover {
-  background-color: #333; 
-}
-
-
-/* Main container styling */
-.Waiter-container {
+/* General Page Background */
+.waiter-container {
   display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  padding: 20px;
+  flex-direction: column;
+  align-items: center;
+  background-color: white; /* White background for the entire page */
+  color: black; /* Black text color */
+  min-height: 100vh;
 }
 
-/* Finalized order table styling */
-.Finalized {
-  width: 28%;
-  margin-right: 20px;
-  padding: 20px;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
-  background-color: #f8f9fa;
-}
-
-.Finalized h3 {
-  margin-bottom: 20px;
-  text-align: center;
-  font-family: 'Ubuntu', sans-serif; /* Use Ubuntu font */
-  font-weight: 600; /* Bold font for emphasis */
-}
-
-/* Finalized order table */
-.final-table {
-  
+/* Navbar styling */
+.navbar {
   width: 100%;
-  border-collapse: collapse;
-}
-
-.final-table th,
-.final-table td {
-  padding: 10px;
-  text-align: left;
-  border-bottom: 1px solid #ddd;
-}
-
-.final-table th {
-  font-weight: bold;
-  color: #333;
-  font-family: 'Ubuntu', sans-serif; /* Use Ubuntu font */
-}
-
-.final-table td {
-  color: #555;
-  font-family: 'Ubuntu', sans-serif; /* Use Ubuntu font */
-}
-
-/* Cards container styling */
-.Cards-container {
-  width: 70%;
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  gap: 20px;
-}
-
-.Card {
-  border-radius: 8px;
-  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
-  text-align: center;
-}
-
-.Card-Body {
-  display: block;
+  background-color: #f8f9fa; /* Light background for the navbar */
+  color: black; /* Black text color in the navbar */
   padding: 15px;
 }
 
-.Card:hover{
-  background: rgb(252, 66, 66);
-  border-radius: 8px;
+.navbar .navbar-brand {
+  font-size: 2rem;
+  font-weight: 700;
+  color: black; /* Ensure navbar text is black */
 }
 
-.Card-Title {
-  font-size: 18px;
-  font-weight: bold;
-  color: #333;
-  font-family: 'Ubuntu', sans-serif;
-}
-
-.Card-Text {
-  display: block;
-  align-items: center;
+/* Content Area Styling */
+.content {
+  display: flex;
+  width: 100%;
+  padding: 20px;
   justify-content: space-between;
 }
 
-.m-cost {
-  font-weight: bold;
-  color: #007bff;
+/* Left and Right Side Sections (65% and 35%) */
+.left-side {
+  flex: 0 0 65%;  /* Left side takes 65% of the space */
+  padding: 20px;
 }
 
-/* Button styling */
-.Card button {
-  margin: 0 5px;
+.right-side {
+  flex: 0 0 35%;  /* Right side takes 35% of the space */
+  padding: 20px;
 }
 
-/* Finalize button styling */
+/* Section Background Styling */
+.finalized-orders, .menu {
+  background-color: #f1f1f1; /* Light background for sections */
+  padding: 15px;
+  border-radius: 10px;
+  margin-bottom: 20px;
+  position: relative; /* Ensure finalize button stays within the section */
+}
+
+/* Section Titles */
+.finalized-orders h3, .menu h3 {
+  color: black; /* Black color for section headers */
+  margin-bottom: 15px;
+}
+
+/* Table for Orders */
+.final-table {
+  width: 100%;
+  border-collapse: collapse;
+  margin-top: 15px;
+}
+
+.final-table th, .final-table td {
+  padding: 10px;
+  border: 1px solid #ccc; /* Light border for table */
+  text-align: center;
+}
+
+.final-table th {
+  background-color: #f8f9fa; /* Light background for table headers */
+}
+
+/* Item Cards Styling */
+.item-cards {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 10px;
+  align-items: flex-start;
+}
+
+
+.item-card {
+  background-color: #ffffff; /* White background for item cards */
+  color: black; /* Black text for item cards */
+  border-radius: 8px;
+  width: 200px;
+  text-align: center;
+  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1); /* Subtle shadow for card effect */
+}
+
+.item-card button {
+  margin: 5px;
+}
+
+/* Finalize Button Styling (Bottom corner of the menu box) */
 .finalize {
-  display: block;
-  margin: 20px auto;
-  padding: 10px 20px;
-  font-size: 16px;
-  font-weight: bold;
-  background-color: #28a745;
-  border-color: #28a745;
-  color: #fff;
-  border-radius: 5px;
-  transition: background-color 0.3s;
-  font-family: 'Ubuntu', sans-serif; /* Use Ubuntu font */
+  position: absolute;
+  bottom: 20px;  /* Positioned 20px from the bottom of the menu box */
+  right: 20px;   /* Positioned 20px from the right of the menu box */
+  background-color: #007bff;
+  color: white;
+  padding: 15px;
+  font-size: 1rem;
+  border-radius: 8px;
+  box-shadow: 0px 2px 10px rgba(0, 0, 0, 0.2);
+  cursor: pointer;
 }
 
 .finalize:hover {
-  background-color: #a2a500;
-  border-color: #ff0000;
+  background-color: #0056b3;
+}
+
+/* Section Visibility Control */
+.finalized-orders.visible,
+.menu.visible {
+  display: block;
+}
+
+.finalized-orders.hidden,
+.menu.hidden {
+  display: none;
+}
+
+/* Icon Button Styling */
+.finalize .icon-button {
+  background: transparent;
+  border: none;
+  color: white;
+  font-size: 1.2rem;
+  cursor: pointer;
+}
+
+/* When Item Card is Selected */
+.item-card.selected {
+  background-color: #d3ffd3; /* Light green background for selected item */
+  border: 2px solid #28a745;
+}
+
+/* Responsive Design for Smaller Screens */
+@media (max-width: 768px) {
+  .content {
+    flex-direction: column;
+  }
+
+  .item-cards {
+    flex-direction: column;
+    gap: 15px;
+  }
+}
+
+@media (max-width: 600px) {
+  .navbar .navbar-brand {
+    font-size: 1.5rem;
+  }
+
+  .item-card {
+    width: 100%;
+  }
 }
