+++ webapp/src/UsePages/Waiter.jsx
@@ -2,11 +2,11 @@ import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Button, Card, Modal, Form } from 'react-bootstrap';
 import axios from 'axios';
-import waitericon from '../Assets/waiter.png';
 import './Waiter.css';
+import titleImage from '../Assets/Zen&Zest.png';
 
 const Waiter = () => {
-  const navigate=useNavigate();
+  const navigate = useNavigate();
   const items = [
     { id: 1, name: 'Item 1', price: 120 },
     { id: 2, name: 'Item 2', price: 150 },
@@ -38,11 +38,10 @@ const Waiter = () => {
       const indexToBeDeleted = newSelectedItems.findIndex(cursor => cursor.name === items[index].name
         && cursor.quantity === (Qty[index] + 1)
         && cursor.price === items[index].price);
-      console.log('DeletedIndex: ' + indexToBeDeleted)
       newSelectedItems.splice(indexToBeDeleted, 1);
       if (Qty[index] === 0) return [...newSelectedItems];
       else return [...newSelectedItems, { name: items[index].name, quantity: Qty[index], price: items[index].price }];
-    })
+    });
   }
 
   function increment(index) {
@@ -58,17 +57,14 @@ const Waiter = () => {
         const indexToBeDeleted = newSelectedItems.findIndex(cursor => cursor.name === items[index].name
           && cursor.quantity === (Qty[index] - 1)
           && cursor.price === items[index].price);
-        console.log('DeletedIndex: ' + indexToBeDeleted)
         newSelectedItems.splice(indexToBeDeleted, 1);
         return [...newSelectedItems, { name: items[index].name, quantity: Qty[index], price: items[index].price }];
       }
-    })
+    });
   }
 
   function handleFinalize() {
     const totalPrice = [...itemsSelected].reduce((accummulator, curr) => accummulator + (curr.quantity * curr.price), 0);
-    /*The above works, reduce is built in and used to iterate through every element in an array, 
-    acc: accummulate, cur: current, 0 initialises acc to 0.*/
     if (totalPrice <= 0) return alert("Please select items before finalizing order.");
     setShowModal(true);
   }
@@ -77,7 +73,6 @@ const Waiter = () => {
     if (!customerName.trim()) return alert("Please enter a valid name.");
 
     try {
-      // Create the order object
       const orderData = {
         name: customerName,
         items: itemsSelected.map(item => ({
@@ -88,12 +83,9 @@ const Waiter = () => {
         totalPrice: itemsSelected.reduce((acc, curr) => acc + (curr.quantity * curr.price), 0),
         status: "Pending"
       };
-      console.log(orderData);
-      // Send POST request to the backend API
       const response = await axios.post(`http://${process.env.REACT_APP_BACKEND_URI}/item`, { ...orderData });
 
-      if (/^20\d$/.test(response.status.toString())) {  // Checking of the response code is in the 200 series
-        // If order is successfully saved, reset quantities and selected items
+      if (/^20\d$/.test(response.status.toString())) {
         setQty(Array(items.length).fill(0));
         setItemsSelected([]);
         setCustomerName('');
@@ -109,117 +101,120 @@ const Waiter = () => {
   async function getJobs() {
     try {
       const response = await axios.get(`http://${process.env.REACT_APP_BACKEND_URI}/items`);
-      let tempQ = [...response.data].filter(data => data.status !== 'Delivered')
+      let tempQ = [...response.data].filter(data => data.status !== 'Delivered');
       setStatusQueue(tempQ);
-      console.log(response.data);
-      return tempQ;
     } catch (error) {
       console.error("Error fetching jobs:", error);
     }
   }
 
-  // Used to debug the crap out of this page
   useEffect(() => {
-    console.log(itemsSelected);
-    console.log('qty: ' + Qty);
     getJobs();
-    return () => {
-      // console.log('decompiler called')
-    }
   }, [itemsSelected, Qty, reloadFlag]);
 
   return (
-    <div className='Header'>
-      <h1>Waiter Page<img src={waitericon} alt="Waiter icon" className="icon" /></h1>
-    <div className="Waiter-container">
-      <div className="Finalized">
+    <div className="waiter-container">
+  <nav className="navbar navbar-light bg-light">
+  <a className="navbar-brand" onClick={() => navigate('/')}>
+    <img src={titleImage} width="200" height="80"/>
+  </a>
+</nav>
+
+  <div className="content">
+    {/* Left section - Items Menu */}
+    <div className="left-side">
+      <div className="menu">
+        <h3>Menu</h3>
+        <Button className="finalize" onClick={handleFinalize}>
+          Finalize Order
+        </Button>
+        <div className="item-cards">
+          {items.map((item, index) => (
+            <Card className="item-card" key={item.id}>
+              <Card.Body>
+                <Card.Title>{item.name}</Card.Title>
+                <Card.Text>
+                  <div>₹{item.price}</div>
+                  <Button onClick={() => decrement(index)} variant="primary" size="sm">-</Button>
+                  <span>{Qty[index]}</span>
+                  <Button onClick={() => increment(index)} variant="secondary" size="sm">+</Button>
+                </Card.Text>
+              </Card.Body>
+            </Card>
+          ))}
+        </div>
+      </div>
+    </div>
+
+    {/* Right section - Booked Orders */}
+    <div className="right-side">
+      <div className="finalized-orders">
         <h3>Booked Orders</h3>
         <table className="final-table">
           <thead>
-              <tr>
+            <tr>
               <th>Customer Name</th>
               <th>Total Price</th>
               <th>Status</th>
             </tr>
           </thead>
           <tbody>
-              {statusQueue.map((order, index) => (
-                <tr key={order._id}>
+            {statusQueue.map(order => (
+              <tr key={order._id}>
                 <td>{order.name}</td>
                 <td>₹{order.totalPrice}</td>
                 <td>
-                    <Button onClick={() => navigate('/order')}>{order.status}</Button>
+                  <Button onClick={() => navigate('/order')}>{order.status}</Button>
                 </td>
               </tr>
-              ))}
+            ))}
           </tbody>
         </table>
       </div>
+    </div>
+  </div>
 
-      <div className="Cards-container">
-        {items.map((item, index) => (
-          <Card className="Card" key={item.id}>
-            <Card.Body className="Card-Body">
-              <Card.Title className="Card-Title">{item.name}</Card.Title>
-              <Card.Text className="Card-Text">
-                <div className="m-cost">₹{item.price}</div>
-                <Button className="Dec" onClick={() => decrement(index)} variant="primary" size="sm">
-                  -
-                </Button>
-                <span>{Qty[index]}</span>
-                <Button className="Inc" onClick={() => increment(index)} variant="secondary" size="sm">
-                  +
-                </Button>
-              </Card.Text>
-            </Card.Body>
-          </Card>
-        ))}
+  {/* Modal for customer details */}
+  <Modal show={showModal} onHide={() => setShowModal(false)}>
+    <Modal.Header closeButton>
+      <Modal.Title>Customer Details</Modal.Title>
+    </Modal.Header>
+    <Modal.Body>
+      <Form>
+        <Form.Group className="mb-3">
+          <Form.Label>Customer Name</Form.Label>
+          <Form.Control
+            type="text"
+            placeholder="Enter customer name"
+            value={customerName}
+            onChange={(e) => setCustomerName(e.target.value)}
+            autoFocus
+          />
+        </Form.Group>
+      </Form>
+      <div>
+        <h5>Order Summary:</h5>
+        <ul>
+          {itemsSelected.map((item, idx) => (
+            <li key={idx}>
+              {item.name} x {item.quantity} = ₹{item.quantity * item.price}
+            </li>
+          ))}
+        </ul>
+        <p><strong>Total: ₹{itemsSelected.reduce((acc, item) => acc + item.quantity * item.price, 0)}</strong></p>
       </div>
-
-        <Button className="finalize" onClick={handleFinalize}>
-        Finalize
+    </Modal.Body>
+    <Modal.Footer>
+      <Button variant="secondary" onClick={() => setShowModal(false)}>
+        Cancel
       </Button>
-    </div>
-      {showModal &&
-        <Modal show={showModal} onHide={() => setShowModal(false)}>
-          <Modal.Header closeButton>
-            <Modal.Title>Customer Details</Modal.Title>
-          </Modal.Header>
-          <Modal.Body>
-            <Form>
-              <Form.Group className="mb-3">
-                <Form.Label>Customer Name</Form.Label>
-                <Form.Control
-                  type="text"
-                  placeholder="Enter customer name"
-                  value={customerName}
-                  onChange={(e) => setCustomerName(e.target.value)}
-                  autoFocus
-                />
-              </Form.Group>
-            </Form>
-            <div>
-              <h5>Order Summary:</h5>
-              <ul>
-                {itemsSelected.map((item, index) => (
-                  <li key={index}>
-                    {item.name} x {item.quantity} = ₹{item.quantity * item.price}
-                  </li>
-                ))}
-              </ul>
-              <p><strong>Total: ₹{itemsSelected.reduce((acc, curr) => acc + curr.quantity * curr.price, 0)}</strong></p>
-            </div>
-          </Modal.Body>
-          <Modal.Footer>
-            <Button variant="secondary" onClick={() => setShowModal(false)}>
-              Cancel
-            </Button>
-            <Button variant="primary" onClick={submitOrder}>
-              Place Order
-            </Button>
-          </Modal.Footer>
-        </Modal>}
-    </div>
+      <Button variant="primary" onClick={submitOrder}>
+        Place Order
+      </Button>
+    </Modal.Footer>
+  </Modal>
+</div>
+
   );
 };
